made some changes after what kokchun feedback
