+++ src/Producer/coin_producer_DOT.py
@@ -1,6 +1,5 @@
 import time
 from quixstreams import Application  
-import json
 from connect_api import get_latest_coin_data   
 
 TARGET_COIN = "DOT"
