+++ src/streamlit_dash/rates_api.py
@@ -1,6 +1,6 @@
-from requests import Session, Timeout, TooManyRedirects
+from requests import Session
 import requests  
-import json
+
